Move Notation support into its own file
